+++ CifraShop.Domain/Entities/User.cs
@@ -11,7 +11,7 @@ namespace CifraShop.Domain.Entities
         public int Id { get; set; }
         public string Email {  get; set; }
         public string Password { get; set; }
-        public decimal ? Balance { get; set; }
+        public short? Balance { get; set; }
         public UserRole Role { get; set; }
         //public ICollection Orders {  get; set; } = new List<Order>();
         public List<Order> Orders { get; set; } = new List<Order>();
